+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ## 3.77.0 (Unreleased)
+
+### Added
+- Support resource discovery for `streaming` resources
+- Support resource discovery for `healthChecks` resources
+- Support resource discovery for `events` resources
+
+### Fixed
+- Fixed DNS outage causing problems for DNS records
+- Fixed string values to escape TF characters in resource discovery
+
 ## 3.76.0 (May 19, 2020)
 
 ### Added